Handle clear-on-read bits for QSFPs
Bits from the same QSFP were being read indiscriminately, but some bits
are clear-on-read, which would get cleared from reading other bits on
the same address.

With this change, QSFP bit operations are tied with other bits of the
same address, so clear-on-read bits are stored and read/cleared
appropriately.

Change-Id: Ied3effa6d56f9373cf814aa43ee6371765f33ed6