Autodetect platforms that need to be loaded
Instead of listing all the platforms one by one, dynamically detect
all the platforms to be loaded. It also has the benefit of reducing
the risk of conflicts when adding multiple platforms.

In the future this will be leveraged by the lazy loading mechanism.
It will only import the current platform definition rather than all of
them, reducing initialisation cost.

Change-Id: I1f4206d9338cc3d27231e758ebeeef1bf84dc7f0