Make older sku detection more reliable
In some cases the cpu smbus are initialized differently by the kernel
leading to a failure to read the eeprom. We now test other address on
which the eeprom could end up.

Both the 7050QX-32 and 7050QX-32S can now be detected via the cmdline
and do not need the eeprom to properly initialize.

Change-Id: I58e43f3ecd0a8ac67ad22d524a86afd7aeed7724