Last refactor of the argument parser
The argument parser is now represented by a tree of Parser nodes.
This tree is built when loading the various parsers under
arista.cli.args. Then when the arguments are processed the associated
action is loaded.

Given the tree structure, we can now have recursive subparsers and have
the parent parsers do some logic and populate a context that the
downstream parser can then use.

Change-Id: Idabc38450bc3c3bec309e347f9985199da513095