Add LazyInventory and MetaInventory classes
MetaInventory is a proxy object to a collection of Inventory objects.
LazyInventory is a lightweight Inventory that only creates the necessary
attribute based on a cache miss mechanism.

Components now all have a LazyInventory thet they should start populate.
FixedSystems still use the original Inventory for now.

Once all components populate their LazyInventory the FixedSystem
inventory will be changed to a MetaInventory.
It will take the iterInventory generator as input.

Change-Id: I46c1d40d981222daedd623c10dd2673ff161b279