Refresh xcvr creation to use optoe driver
The sff8436 driver has been superseeded by the optoe one.
A legacy API allowed us to continue creating the devices as before.
However SFP xcvrs are often present at 2 or 3 addresses on the i2c bus
which the optoe kernel module can handle.
In the case of SFP we now use optoe2 which will manage 0x50 and 0x51.

Change-Id: I542bee68be085df5b23f8c11b59c72a36ff0dcc0