Implement Sfp methods that use sfputil and fix enumeration
There are some functionalities in xcvrd that is required of these Sfp
classes. We can have them call into sfputil for this for now. This
solves an issue in xcvrd that will lead it to crash, since SfpUtilHelper,
the back-up class, does not implement the required functions.

Arista enumerates index of xcvrs not to not start at 0. Typically it
starts at 1 but there are some exceptions such as 7050QX-32S.
This change will insert blank entries so xcvrd methods will fetch the
right Sfp objects.

Change-Id: I8a87a65156fb1eb56ac78dd4778cebe3c242ed34